simplify webview view and panel
